add 1750

+++ LeetCode/1750_删除字符串两端相同字符后的最短长度.cpp
@@ -18,6 +18,6 @@ public:
                 j--;
             }
         }
-        return j - i + 1;
+        return static_cast<int>(j - i + 1);
     }
 };
